user flow code refactor
