+++ quora-api/src/test/java/com/upgrad/quora/api/controller/AdminControllerTest.java
@@ -34,7 +34,7 @@ public class AdminControllerTest {
     //This test case passes when you try to delete the user but the role of the user corresponding to the JWT token entered is nonadmin.
     @Test
     public void deleteWithnonadminAsRole() throws Exception {
-        mvc.perform(MockMvcRequestBuilders.delete("/admin/user/database_uuid1").header("authorization", "Bearer database_accesstoken1"))
+        mvc.perform(MockMvcRequestBuilders.delete("/admin/user/database_uuid4").header("authorization", "database_accesstoken1"))
                 .andExpect(status().isForbidden())
                 .andExpect(MockMvcResultMatchers.jsonPath("code").value("ATHR-003"));
     }
@@ -43,7 +43,7 @@ public class AdminControllerTest {
     //This test case passes when you try to delete the user which does not exist in the database.
     @Test
     public void deleteNonExistingUser() throws Exception {
-        mvc.perform(MockMvcRequestBuilders.delete("/admin/user/non_existing_user_uuid").header("authorization", "Bearer database_accesstoken"))
+        mvc.perform(MockMvcRequestBuilders.delete("/admin/user/non_existing_user_uuid").header("authorization", "database_accesstoken"))
                 .andExpect(status().isNotFound())
                 .andExpect(MockMvcResultMatchers.jsonPath("code").value("USR-001"));
     }
