+++ quora-service/src/main/java/com/upgrad/quora/service/business/UserBussinessService.java
@@ -5,6 +5,9 @@ import com.upgrad.quora.service.dao.UserDao;
 import com.upgrad.quora.service.entity.UserAuthTokenEntity;
 import com.upgrad.quora.service.entity.UserEntity;
 import com.upgrad.quora.service.exception.AuthenticationFailedException;
+import com.upgrad.quora.service.exception.AuthorizationFailedException;
+import com.upgrad.quora.service.exception.SignOutRestrictedException;
+import com.upgrad.quora.service.exception.SignUpRestrictedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -13,11 +16,10 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.ZonedDateTime;
 import java.util.Base64;
 
-import static com.upgrad.quora.service.common.GenericErrorCode.ATH_001;
-import static com.upgrad.quora.service.common.GenericErrorCode.ATH_002;
+import static com.upgrad.quora.service.common.GenericErrorCode.*;
 
 @Service
-public class SigninAuthenticationService {
+public class UserBussinessService {
 
     @Autowired
     private UserDao userDao;
@@ -25,6 +27,15 @@ public class SigninAuthenticationService {
     @Autowired
     private PasswordCryptographyProvider cryptographyProvider;
 
+    @Transactional(propagation = Propagation.REQUIRED)
+    public UserEntity signup(final UserEntity userEntity) throws SignUpRestrictedException {
+        validateUserDetails(userEntity);
+        String[] encryptedText = cryptographyProvider.encrypt(userEntity.getPassword());
+        userEntity.setSalt(encryptedText[0]);
+        userEntity.setPassword(encryptedText[1]);
+        return userDao.createUser(userEntity);
+
+    }
 
     @Transactional(propagation = Propagation.REQUIRED)
     public UserAuthTokenEntity authenticate(final String authorization) throws AuthenticationFailedException {
@@ -65,4 +76,43 @@ public class SigninAuthenticationService {
         }
 
     }
+
+    @Transactional(propagation = Propagation.REQUIRED)
+    public UserEntity signoutUser(final String authorization) throws SignOutRestrictedException {
+        UserAuthTokenEntity userAuthToken = getUserAuthToken(authorization);
+        if (userAuthToken.getLogoutAt() != null) {
+            throw new SignOutRestrictedException(SGOR_001.getCode(), SGOR_001.getDefaultMessage());
+        }
+        userAuthToken.setLogoutAt(ZonedDateTime.now());
+        userAuthToken = userDao.updateAuthToken(userAuthToken);
+
+        return userAuthToken.getUser();
+    }
+
+
+    private UserAuthTokenEntity getUserAuthToken(final String authorizationToken) throws
+            SignOutRestrictedException {
+        UserAuthTokenEntity userAuthTokenEntity = userDao.getUserAuthToken(authorizationToken);
+        if (userAuthTokenEntity == null) {
+            throw new SignOutRestrictedException(SGOR_001.getCode(), SGOR_001.getDefaultMessage());
+        }
+        return userAuthTokenEntity;
+    }
+
+
+    private void validateUserDetails(final UserEntity user) throws
+            SignUpRestrictedException {
+
+        UserEntity userEntity = userDao.getUserByUserName(user.getUserName());
+        if (userEntity != null) {
+            throw new SignUpRestrictedException(SGUR_001.getCode(), SGUR_001.getDefaultMessage());
+        }
+
+        userEntity = userDao.getUserByEmail(user.getEmailAddress());
+        if (userEntity != null) {
+            throw new SignUpRestrictedException(SGUR_002.getCode(), SGUR_002.getDefaultMessage());
+        }
+
+    }
 }
+
